janitor: Run `cargo clippy --fix`
Only the hand-picked sensible things, not all of it ;-)

Also fix a few typos that cspell complained about when I
tried to commit and some formatting changes that cargo fmt
insisted on when commiting.

+++ internal/core/items/text.rs
@@ -1082,7 +1082,7 @@ impl TextInput {
         self.anchor_position_byte_offset.set(anchor as i32);
 
         self.add_undo_item(UndoItem {
-            pos: anchor as usize,
+            pos: anchor,
             text: removed_text,
             cursor: real_cursor,
             anchor: real_anchor,
@@ -1183,7 +1183,7 @@ impl TextInput {
         }
 
         self.add_undo_item(UndoItem {
-            pos: cursor_pos as usize,
+            pos: cursor_pos,
             text: inserted_text,
             cursor: real_cursor,
             anchor: real_anchor,
